Adding StatisticEventProcessor and modify Event classes accordingly
