+++ modules/core/src/main/java/org/apache/synapse/aspects/flow/statistics/log/templates/CallbackHandledEvent.java
@@ -19,8 +19,8 @@
 package org.apache.synapse.aspects.flow.statistics.log.templates;
 
 import org.apache.synapse.aspects.flow.statistics.collectors.CallbackStatisticCollector;
-import org.apache.synapse.aspects.flow.statistics.collectors.RuntimeStatisticCollector;
 import org.apache.synapse.aspects.flow.statistics.data.raw.CallbackDataUnit;
+import org.apache.synapse.aspects.flow.statistics.log.StatisticEventProcessor;
 import org.apache.synapse.aspects.flow.statistics.log.StatisticsReportingEvent;
 import org.apache.synapse.aspects.flow.statistics.util.StatisticsConstants;
 
@@ -37,7 +37,7 @@ public class CallbackHandledEvent implements StatisticsReportingEvent {
 
 	@Override
 	public void process() {
-		CallbackStatisticCollector.removeCallback(callbackDataUnit);
-		RuntimeStatisticCollector.closeStatisticEntry(callbackDataUnit, StatisticsConstants.ATTEMPT_TO_CLOSE);
+		StatisticEventProcessor.removeCallback(callbackDataUnit);
+		StatisticEventProcessor.closeStatisticEntry(callbackDataUnit, StatisticsConstants.ATTEMPT_TO_CLOSE);
 	}
 }
