+++ modules/core/src/main/java/org/apache/synapse/aspects/flow/statistics/log/templates/CallbackSentEvent.java
@@ -18,8 +18,8 @@
 
 package org.apache.synapse.aspects.flow.statistics.log.templates;
 
-import org.apache.synapse.aspects.flow.statistics.collectors.CallbackStatisticCollector;
 import org.apache.synapse.aspects.flow.statistics.data.raw.CallbackDataUnit;
+import org.apache.synapse.aspects.flow.statistics.log.StatisticEventProcessor;
 import org.apache.synapse.aspects.flow.statistics.log.StatisticsReportingEvent;
 
 /**
@@ -35,6 +35,6 @@ public class CallbackSentEvent implements StatisticsReportingEvent {
 
 	@Override
 	public void process() {
-		CallbackStatisticCollector.addCallbacks(callbackDataUnit);
+		StatisticEventProcessor.addCallbacks(callbackDataUnit);
 	}
 }
