+++ modules/core/src/main/java/org/apache/synapse/aspects/flow/statistics/log/templates/FinalizedFlowEvent.java
@@ -18,8 +18,8 @@
 
 package org.apache.synapse.aspects.flow.statistics.log.templates;
 
-import org.apache.synapse.aspects.flow.statistics.collectors.RuntimeStatisticCollector;
 import org.apache.synapse.aspects.flow.statistics.data.raw.BasicStatisticDataUnit;
+import org.apache.synapse.aspects.flow.statistics.log.StatisticEventProcessor;
 import org.apache.synapse.aspects.flow.statistics.log.StatisticsReportingEvent;
 import org.apache.synapse.aspects.flow.statistics.util.StatisticsConstants;
 
@@ -36,6 +36,6 @@ public class FinalizedFlowEvent implements StatisticsReportingEvent {
 
 	@Override
 	public void process() {
-		RuntimeStatisticCollector.closeStatisticEntry(basicStatisticDataUnit, StatisticsConstants.ATTEMPT_TO_CLOSE);
+		StatisticEventProcessor.closeStatisticEntry(basicStatisticDataUnit, StatisticsConstants.ATTEMPT_TO_CLOSE);
 	}
 }
