+++ modules/core/src/main/java/org/apache/synapse/aspects/flow/statistics/log/templates/StatisticsOpenEvent.java
@@ -18,8 +18,8 @@
 
 package org.apache.synapse.aspects.flow.statistics.log.templates;
 
-import org.apache.synapse.aspects.flow.statistics.collectors.RuntimeStatisticCollector;
 import org.apache.synapse.aspects.flow.statistics.data.raw.StatisticDataUnit;
+import org.apache.synapse.aspects.flow.statistics.log.StatisticEventProcessor;
 import org.apache.synapse.aspects.flow.statistics.log.StatisticsReportingEvent;
 
 /**
@@ -35,6 +35,6 @@ public class StatisticsOpenEvent implements StatisticsReportingEvent {
 
 	@Override
 	public void process() {
-		RuntimeStatisticCollector.recordStatisticsOpenEvent(statisticDataUnit);
+		StatisticEventProcessor.openStatisticEntry(statisticDataUnit);
 	}
 }
